Refactor main; fix tests
