+++ examples/client_pool/Cargo.toml
@@ -10,3 +10,5 @@ path = "client_pool.rs"
 [dependencies]
 lychee-lib = { path = "../../lychee-lib", version = "0.7.2" }
 tokio = { version = "1.12.0", features = ["full"] }
+futures = "0.3.17"
+tokio-stream = "0.1.7"
