+++ examples/client_pool/client_pool.rs
@@ -1,14 +1,14 @@
-use lychee_lib::{ClientBuilder, ClientPool, Input, Request, Result, Uri};
+use lychee_lib::{ClientBuilder, Input, Request, Result, Uri};
 use std::convert::TryFrom;
 use tokio::sync::mpsc;
+use tokio_stream::wrappers::ReceiverStream;
 
 const CONCURRENT_REQUESTS: usize = 4;
 
 #[tokio::main]
-#[allow(clippy::trivial_regex)]
 async fn main() -> Result<()> {
     // These channels are used to send requests and receive responses to and
-    // from the lychee client pool
+    // from lychee
     let (send_req, recv_req) = mpsc::channel(CONCURRENT_REQUESTS);
     let (send_resp, mut recv_resp) = mpsc::channel(CONCURRENT_REQUESTS);
 
@@ -18,7 +18,7 @@ async fn main() -> Result<()> {
         Input::Stdin,
     )];
 
-    // Send requests to pool
+    // Queue requests
     tokio::spawn(async move {
         for request in requests {
             println!("Sending {}", request);
@@ -29,13 +29,18 @@ async fn main() -> Result<()> {
     // Create a default lychee client
     let client = ClientBuilder::default().client()?;
 
-    // Create a pool with four lychee clients
-    let clients = vec![client; CONCURRENT_REQUESTS];
-    let mut clients = ClientPool::new(send_resp, recv_req, clients);
-
-    // Handle requests in a client pool
+    // Start receiving requests
+    // Requests get streamed into the client and run concurrently
     tokio::spawn(async move {
-        clients.listen().await;
+        futures::StreamExt::for_each_concurrent(
+            ReceiverStream::new(recv_req),
+            CONCURRENT_REQUESTS,
+            |req| async {
+                let resp = client.check(req).await.unwrap();
+                send_resp.send(resp).await.unwrap();
+            },
+        )
+        .await;
     });
 
     // Finally, listen to incoming responses from lychee
